Chains moved to coroutines. Cumulatives to core
