+++ kmath-commons/build.gradle.kts
@@ -7,7 +7,7 @@ description = "Commons math binding for kmath"
 
 dependencies {
     api(project(":kmath-core"))
-    api(project(":kmath-sequential"))
+    api(project(":kmath-streaming"))
     api("org.apache.commons:commons-math3:3.6.1")
     testImplementation("org.jetbrains.kotlin:kotlin-test")
     testImplementation("org.jetbrains.kotlin:kotlin-test-junit")
