+++ kmath-core/src/commonTest/kotlin/scientifik/kmath/misc/CumulativeKtTest.kt
@@ -1,6 +1,5 @@
 package scientifik.kmath.misc
 
-import scientifik.kmath.sequential.cumulativeSum
 import kotlin.test.Test
 import kotlin.test.assertEquals
