+++ kmath-coroutines/src/jvmMain/kotlin/scientifik/kmath/chains/ChainExt.kt
@@ -1,6 +1,7 @@
-package scientifik.kmath.sequential
+package scientifik.kmath.chains
 
 import kotlinx.coroutines.runBlocking
+import sicentifik.kmath.chains.Chain
 import kotlin.sequences.Sequence
 
 /**
